+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+* Created/updated at timestamps for database entities
+* Queries for when a volunteer last updated their skills, and which skills have been added/changed since this
+
 ### Changed
 
 * Refactor charts to a standalone stats page